# packages/proto/gen/openapi/rioku/v1/api.full.json
{
  "components": {
    "parameters": {
      "ID": {
        "description": "Resource id (UUID).",
        "in": "path",
        "name": "id",
        "required": true,
        "schema": {
          "format": "uuid",
          "type": "string"
        }
      },
      "Tenant": {
        "description": "Tenant slug (e.g. `default`, `acme`).",
        "in": "path",
        "name": "tenant",
        "required": true,
        "schema": {
          "pattern": "^[a-z0-9-]+$",
          "type": "string"
        }
      }
    },
...
    "securitySchemes": {
      "bearerToken": {
        "bearerFormat": "JWT",
        "description": "API key or short-lived JWT issued by /api/v1/auth/token.",
        "scheme": "bearer",
        "type": "http"
      },
      "sessionCookie": {
        "description": "Web-UI session cookie issued by /api/v1/auth/login.",
        "in": "cookie",
        "name": "rioku_sid",
        "type": "apiKey"
      }
    }
  },
  "consumes": [
    "application/json"
  ],
...
  "info": {
    "contact": {
      "name": "Rioku Labs",
      "url": "https://github.com/riokulabs/rioku"
    },
    "description": "Rioku is an open-source, AI-native API gateway platform built on\nCaddy. This document is the canonical OpenAPI 3.1 specification\nfor every REST endpoint exposed by the daemon.\n\nConventions:\n\n- Tenant-scoped paths live under `/api/v1/t/{tenant}/...`.\n- Every resource exposes `OPTIONS` for capability discovery.\n- Both `PUT` (replace, full body) and `PATCH` (partial, RFC 7396\n  merge-patch) are supported on every resource that takes\n  mutations.\n- Relations carry both the foreign-key id AND a `_links.\u003crelation\u003e.href`\n  pointing at the canonical resource path.\n- Pagination is cursor-based (`?page_token=`); page size is\n  bounded at 1000.\n",
    "license": {
      "name": "BUSL-1.1",
      "url": "https://github.com/riokulabs/rioku/blob/main/LICENSE"
    },
    "title": "Rioku API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/t/{tenant}/routes/{id}": {
      "delete": {
        "operationId": "deleteRoute",
        "responses": {
          "204": {
            "description": "deleted"
          }
        },
        "tags": [
          "routes"
        ]
      },
      "get": {
        "operationId": "getRoute",
        "responses": {
          "200": {
            "description": "ok"
          }
        },
        "tags": [
          "routes"
        ]
      },
      "options": {
        "responses": {
          "204": {
            "description": "no content"
          }
        }
      },
      "parameters": [
        {
          "$ref": "#/components/parameters/Tenant"
        },
        {
          "$ref": "#/components/parameters/ID"
        }
      ],
      "patch": {
        "operationId": "patchRoute",
        "requestBody": {
          "content": {
            "application/merge-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/v1Route"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "ok"
          }
        },
        "tags": [
          "routes"
        ]
      },
      "put": {
        "operationId": "replaceRoute",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/v1Route"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "ok"
          }
        },
        "tags": [
          "routes"
        ]
      }
    },
...
  },
  "produces": [
    "application/json"
  ],
  "security": [
    {
      "sessionCookie": []
    },
    {
      "bearerToken": []
    }
  ],
  "servers": [
    {
      "description": "Rioku daemon REST endpoint",
      "url": "{scheme}://{host}/api/v1",
      "variables": {
        "host": {
          "default": "localhost:7778"
        },
        "scheme": {
          "default": "https",
          "enum": [
            "https",
            "http"
          ]
        }
      }
    }
  ],
  "swagger": "2.0",
...
}